More accurate speeds (not finished)

+++ Game.Engine/Core/Fleet.cs
@@ -17,7 +17,7 @@
         public virtual float ShotThrustB { get => World.Hook.ShotThrustB; }
         public virtual float BaseThrustM { get => World.Hook.BaseThrustM; }
         public virtual float BaseThrustB { get => World.Hook.BaseThrustB; }
-        public virtual int[] BaseThrust { get => World.Hook.BaseThrust; }
+        public virtual float[] BaseThrust { get => World.Hook.BaseThrust; }
         public virtual float BaseThrustConverter { get => World.Hook.BaseThrustConverter; }
         public virtual float BoostThrust { get => World.Hook.BoostThrust; }
